# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: vlxd-v3 API
  version: 0.1.0
  description: Contract for the vlxd-v3 construction-materials management API.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /products:
    get:
      operationId: listProducts
      summary: List products for the authenticated tenant
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: search
          schema: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
        "403":
          description: Permission denied (requires products.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
    post:
      operationId: createProduct
      summary: Create a product for the authenticated tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input or unknown unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
        "403":
          description: Permission denied (requires products.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
        "409":
          description: SKU already exists for the tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
        "422":
          description: Free plan product limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
...
  /stock-receipts:
    get:
      operationId: listStockReceipts
      summary: List stock receipts for the authenticated tenant
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: warehouseId
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of stock receipts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "403":
          description: Permission denied (requires inventory.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
    post:
      operationId: createStockReceipt
      summary: Create an inbound stock receipt and update stock levels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockReceiptRequest"
      responses:
        "201":
          description: Stock receipt created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptDetailResponse"
        "400":
          description: Invalid request payload or lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "403":
          description: Permission denied (requires inventory.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "404":
          description: Warehouse or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
  /stock-receipts/{id}:
    get:
      operationId: getStockReceipt
      summary: Get details of a single stock receipt
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Stock receipt details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptDetailResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "403":
          description: Permission denied (requires inventory.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
        "404":
          description: Stock receipt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReceiptErrorResponse"
  /customers:
    get:
      operationId: listCustomers
      summary: List customers for the authenticated tenant
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
        "403":
          description: Permission denied (requires customers.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
    post:
      operationId: createCustomer
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Invalid customer data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
        "403":
          description: Permission denied (requires customers.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
        "409":
          description: Customer code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerErrorResponse"
  /sales-orders:
    get:
      operationId: listSalesOrders
      summary: List sales orders for the authenticated tenant
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: customerId
          schema: { type: string }
        - in: query
          name: warehouseId
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of sales orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "403":
          description: Permission denied (requires sales.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
    post:
      operationId: createSalesOrder
      summary: Create a sales order and deduct inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesOrderRequest"
      responses:
        "201":
          description: Sales order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderDetailResponse"
        "400":
          description: Invalid request payload or lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "403":
          description: Permission denied (requires sales.create)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "404":
          description: Customer, warehouse, or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "422":
          description: Insufficient stock for one or more products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
  /sales-orders/{id}:
    get:
      operationId: getSalesOrder
      summary: Get details of a single sales order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Sales order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderDetailResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "403":
          description: Permission denied (requires sales.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
        "404":
          description: Sales order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrderErrorResponse"
...
components:
  schemas:
...
    UnitCode:
      type: string
      enum: [vien, bao, tan, kg, m3, cay, tam, thung]
    Product:
      type: object
      additionalProperties: false
      required: [id, sku, name, unitCode, unitName, createdAt, stockLevels]
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        unitCode: { $ref: "#/components/schemas/UnitCode" }
        unitName: { type: string }
        createdAt: { type: string, format: date-time }
        stockLevels:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [warehouseId, warehouseCode, quantity]
            properties:
              warehouseId: { type: string }
              warehouseCode: { type: string }
              quantity: { type: integer, minimum: 0 }
    ProductListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [sku, name, unitCode]
      properties:
        sku: { type: string, minLength: 1, maxLength: 50 }
        name: { type: string, minLength: 1, maxLength: 200 }
        unitCode: { $ref: "#/components/schemas/UnitCode" }
    ProductErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [
              UNAUTHORIZED,
              FORBIDDEN,
              VALIDATION_ERROR,
              PRODUCT_LIMIT_REACHED,
              PRODUCT_SKU_EXISTS,
              UNIT_NOT_FOUND,
            ]
        message: { type: string }
...
    CreateStockReceiptLineInput:
      type: object
      additionalProperties: false
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1, maximum: 1000000 }
    CreateStockReceiptRequest:
      type: object
      additionalProperties: false
      required: [warehouseId, lines]
      properties:
        warehouseId: { type: string }
        note: { type: string, maxLength: 500 }
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateStockReceiptLineInput"
    StockReceiptLine:
      type: object
      additionalProperties: false
      required: [id, productId, productSku, productName, unitName, quantity]
      properties:
        id: { type: string }
        productId: { type: string }
        productSku: { type: string }
        productName: { type: string }
        unitName: { type: string }
        quantity: { type: integer, minimum: 1 }
    StockReceiptListItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          receiptNumber,
          warehouseId,
          warehouseCode,
          warehouseName,
          status,
          createdByName,
          createdAt,
          itemCount,
          totalQuantity,
        ]
      properties:
        id: { type: string }
        receiptNumber: { type: string }
        warehouseId: { type: string }
        warehouseCode: { type: string }
        warehouseName: { type: string }
        status: { type: string }
        note:
          type: [string, "null"]
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
        itemCount: { type: integer }
        totalQuantity: { type: integer }
    StockReceiptListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockReceiptListItem"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    StockReceiptDetailResponse:
      type: object
      additionalProperties: false
      required:
        [
          id,
          receiptNumber,
          warehouseId,
          warehouseCode,
          warehouseName,
          status,
          createdByName,
          createdAt,
          lines,
          totalQuantity,
        ]
      properties:
        id: { type: string }
        receiptNumber: { type: string }
        warehouseId: { type: string }
        warehouseCode: { type: string }
        warehouseName: { type: string }
        status: { type: string }
        note:
          type: [string, "null"]
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
        totalQuantity: { type: integer }
        lines:
          type: array
          items:
            $ref: "#/components/schemas/StockReceiptLine"
    StockReceiptErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [UNAUTHORIZED, FORBIDDEN, WAREHOUSE_NOT_FOUND, PRODUCT_NOT_FOUND, INVALID_RECEIPT_LINES]
        message: { type: string }
    Customer:
      type: object
      additionalProperties: false
      required: [id, code, name, createdAt]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        phone:
          type: [string, "null"]
        address:
          type: [string, "null"]
        createdAt: { type: string, format: date-time }
    CreateCustomerRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        phone:
          type: string
        address:
          type: string
    CustomerListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total: { type: integer }
    CustomerErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [UNAUTHORIZED, FORBIDDEN, VALIDATION_ERROR, CUSTOMER_CODE_EXISTS, INVALID_CUSTOMER_DATA]
        message: { type: string }
    CreateSalesOrderLineInput:
      type: object
      additionalProperties: false
      required: [productId, quantity, unitPrice]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1, maximum: 1000000 }
        unitPrice: { type: integer, minimum: 0, maximum: 100000000000 }
    CreateSalesOrderRequest:
      type: object
      additionalProperties: false
      required: [customerId, warehouseId, lines]
      properties:
        customerId: { type: string }
        warehouseId: { type: string }
        note: { type: string }
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateSalesOrderLineInput"
    SalesOrderLine:
      type: object
      additionalProperties: false
      required: [id, productId, productSku, productName, unitName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: string }
        productId: { type: string }
        productSku: { type: string }
        productName: { type: string }
        unitName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: integer }
        lineTotal: { type: integer }
    SalesOrderListItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          orderNumber,
          customerId,
          customerName,
          warehouseId,
          warehouseName,
          status,
          totalAmount,
          itemCount,
          createdByName,
          createdAt,
        ]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        warehouseId: { type: string }
        warehouseName: { type: string }
        status: { type: string }
        totalAmount: { type: integer }
        itemCount: { type: integer }
        note:
          type: [string, "null"]
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
    SalesOrderListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderListItem"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    SalesOrderDetailResponse:
      type: object
      additionalProperties: false
      required:
        [
          id,
          orderNumber,
          customerId,
          customerCode,
          customerName,
          warehouseId,
          warehouseCode,
          warehouseName,
          status,
          totalAmount,
          createdByName,
          createdAt,
          lines,
        ]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        customerId: { type: string }
        customerCode: { type: string }
        customerName: { type: string }
        customerPhone:
          type: [string, "null"]
        customerAddress:
          type: [string, "null"]
        warehouseId: { type: string }
        warehouseCode: { type: string }
        warehouseName: { type: string }
        status: { type: string }
        totalAmount: { type: integer }
        note:
          type: [string, "null"]
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderLine"
    SalesOrderErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [
              UNAUTHORIZED,
              FORBIDDEN,
              ORDER_NOT_FOUND,
              CUSTOMER_NOT_FOUND,
              WAREHOUSE_NOT_FOUND,
              PRODUCT_NOT_FOUND,
              INSUFFICIENT_STOCK,
              INVALID_ORDER_LINES,
            ]
        message: { type: string }
...